+++ framework/domain-db/src/main/java/is/codion/framework/domain/db/MetaDataSchema.java
@@ -16,9 +16,9 @@
  *
  * Copyright (c) 2020 - 2024, Björn Darri Sigurðsson.
  */
-package is.codion.tools.generator.domain;
+package is.codion.framework.domain.db;
 
-import is.codion.tools.generator.domain.MetaDataTable.TablePacker;
+import is.codion.framework.domain.db.MetaDataTable.TablePacker;
 
 import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;