+++ framework/domain-db/src/main/java/module-info.java
@@ -14,17 +14,15 @@
  * You should have received a copy of the GNU General Public License
  * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
  *
- * Copyright (c) 2023 - 2024, Björn Darri Sigurðsson.
+ * Copyright (c) 2024, Björn Darri Sigurðsson.
  */
 /**
- * Framework domain source generator model classes, such as:<br>
+ * Framework database domain model classes, such as:<br>
  * <br>
- * {@link is.codion.tools.generator.domain.DomainSource}
+ * {@link is.codion.framework.domain.db.DatabaseDomain}
  */
-module is.codion.tools.generator.domain {
-	requires java.compiler;
-	requires com.squareup.javapoet;
-	requires transitive is.codion.framework.domain.db;
+module is.codion.framework.domain.db {
+	requires transitive is.codion.framework.domain;
 
-	exports is.codion.tools.generator.domain;
+	exports is.codion.framework.domain.db;
 }